some code refactoring and optimization
