+++ src/DriftDNS.App/DriftDNS.App.csproj
@@ -3,8 +3,6 @@
   <ItemGroup>
     <ProjectReference Include="..\DriftDNS.Core\DriftDNS.Core.csproj" />
     <ProjectReference Include="..\DriftDNS.Infrastructure\DriftDNS.Infrastructure.csproj" />
-    <ProjectReference Include="..\DriftDNS.Providers.Route53\DriftDNS.Providers.Route53.csproj" />
-    <ProjectReference Include="..\DriftDNS.Providers.Cloudflare\DriftDNS.Providers.Cloudflare.csproj" />
   </ItemGroup>
 
   <ItemGroup>
