+++ tests/DriftDNS.Tests/DriftDNS.Tests.csproj
@@ -20,7 +20,6 @@
   <ItemGroup>
     <ProjectReference Include="..\..\src\DriftDNS.Core\DriftDNS.Core.csproj" />
     <ProjectReference Include="..\..\src\DriftDNS.Infrastructure\DriftDNS.Infrastructure.csproj" />
-    <ProjectReference Include="..\..\src\DriftDNS.Providers.Cloudflare\DriftDNS.Providers.Cloudflare.csproj" />
   </ItemGroup>
 
 </Project>
